Defer checking the rangeOps until the Range call
Per the review comments, this change makes range ops just save
things until resolved later.  The resolution function,
convertRangeOpConditions, does some minimal checking to ensure it
can be sent to an RPC call.

Not included in this diff is the actual implementation of Range()
on the client side, which will use the convertRangeOpConditions.

Other minor changes include improving the error from the registry
and changing the RangeOp to be a pointer.