Change fieldsToRead -> minimumFields
Because the contract is better defined for connectors,
the fieldsToRead name is misleading; it's actually the
minimum fields that the connector must return. It can
return others.

This commit is just a name change; the actual change is
in a different commit. It also does not change the names
in the RPC structures; this is trickier since you have to
change the IDL in tandem.