Add memory connector (#147)
* Add memory connector

The memory connector can be used for testing, or when you just need an in-memory
datastore. This implementation is fairly efficient, mimicking Cassandra's handling of
ordering.

This change also modifies the contract of the connectors: they now may return more data
than originally requested, and expect the callers not to mutate the results. This is safe since
really the only caller should be the Client interface implementation (or tests).

TODO:
Range and Scan always return all the rows, ignoring the specified limit
As a result, you don't get any continuation tokens