Add support for memory connector Range() tokens
This implements continuation tokens for the Range() operator in the
memory connector. Just before returning from Range(), if the length
of the return value is longer than requested, we encode the data
in the last value returned to the user. The encoding is a base64
encoded gob.

When a token is supplied on a future call to Range(), we still find
all the rows, and then decode the token, and look for that value
in the result set. If we find it, then we just start at the next
value. If it's not there, that's okay too; we just start at the
next value in the set. The latter can only happen when the specific
value in the token has been deleted.