Bulk of Cassandra connector work
This shows the differences from the closed source connector and
should be the focal point of the review.

The schema work is in the next diff. The tests do not pass yet because
they depend on the schema being created by the schema diff.

This connector allows for a direct connection to Cassandra from
the client side. To configure the connector, you can either
specify a yaml fragment or a gocql configuration block. The default
is to connect to 127.0.0.1:9042.

Most of the DML code is just ported over from our closed source
connector, with a few changes:
  - Removal of any remaining stapi-go code
  - Removal of any remaining UNS code
  - Changes to test infrastructure to use ccm
  - Renamed Resolver to KeyspaceMapper
  - Abstracted KeyspaceMapper interface for mapping
    (scope, namePrefix) -> keyspace

TODO:
  - Provide a better way to just specify a set of contact points
  - If a generated keyspace name is >48 characters, hash it instead