Ensure Registered Entities are created with Index information (#200)
NewRegisteredEntity has a bug where the entity that get's registered doesn't properly copy over the secondary Indexes for a given EntityDefintion, it only sets the name, key, and columns.

This causes Range lookups on secondary indexes to fail for the in-memory connector.

Let's fix this problem by just using the full EntityDefinition that's available to us via the table that's passed in to the NewRegisteredEntity function. This way, if we ever add any more fields to the EntityDefinition in the future, they'll be automatically included in the RegisteredEntity.