Fallback cache connector (#231)
The fallback cache connector is a connector that reads from an origin connector, and if the origin results in  error, can default to reading from a fallback connector for data. The purpose of the fallback cache connector is to provide a backup solution for reading data in case the origin is having problems. 

The fallback path is only used when reading entities that are designated as cacheable. Otherwise, results from the origin are always returned.

The primary goal of the fallback cache connector is to provide availability to clients.