Fix yarpc tests (#242)
Fix frequent failures of the scan test.

The scan test for the yarpc connector frequently fails.
I believe this is due to the fact that request for the
scan call on the yaprc connector look the same for the
error case and the not found case. This changes those
two mock setups to be distinct.