Adding a TestMain in Cassandra connector package to setup and drop scope once (#252)
We are seeing many test failures with the same error of complaining not able to add an already-existing keyspace. e.g.

Unable to create keyspace "example": Cannot add existing keyspace "example"
This is because in each unit test, we drop the scope and recreate the scope and schema without synchronization, which could run into the above race condition.

Adding a TestMain to setup the Cassandra, create the test scope and schema before tests, and drop the scope after tests finish.