Fix Compound Partition Keys in In-Memory Connector (#281)
Compound Partition Keys are completely broken with the in-memory connector. Each time a value of a partition key is encoded in partitionKeyBuilder the previous encoded partition key value is completely overwritten. This means for entities that have more than one column as part of their partition key, only the last partition key value is used when creating the encoded partition key value.

The result of this is that if two rows have the same value for the last column that's part of their partition key (but different values for the other columns making up their partition key), the second row will completely overwrite the first (since the resulting encoded partition key is the same for both rows).

Let's fix this by concating all of the encoded values for each partition key column and using the result as the encoded partition key.

Let's also write a test that proves compound partition keys now work.