fix bad race conditions in memory connector (#282)
Currently, when reading, ranging, or scanning from the in-memory connector, the connector returns references to internal data (in the forms of slices and maps). This allows callers to read/write internal data without acquiring the appropriate locks inside the in-memory connector, resulting in race conditions.

By returning copies of the values stored within the in-memory connector, we can ensure the values returned to the caller have no references to the internal state within the in-memory connector and prevent data races.