fix upsert race condition (#283)
When upserting (and similarly with CreateIfNotExists) the map that is placed into the in-memory connector is the same reference that the caller also owns. This means that the caller can then make future mutations to that map (which is now considered part of the internal state of the in-memory connector), without acquiring the necessary locks. This causes a race condition.

We can fix this by copying data that is being upserted so the in-memory connector fully owns the data that is part of it's internal state.