Config Refactor (#321)
Refactor the config package

Having the ability to specify which connector you want to use via yaml has an unintended consequence of us needing to import every single connector into the config package. This is brittle (it doesn't really work well with connectors that don't live in this repo) and this means many users of DOSA have to take on glide dependencies for packages they never end up using.

We opt to get rid of the config package as we know it today. However, for users using the yarpc package, we still want to provide a yaml config option. So we create new functions in the yarpc package allowing for the creation of a client based on a specific yarpc config.