Add support for nil partition key for secondary indexes in memory connector
Why:
Cassandra supports secondary indexes where the partition key can be
nil. This results in the index not being generated until the value is
set. The test client which uses a memory connector should support the
same behavior.

-

This change addresses the need by:
Add a `nil` check in `partitionKeyBuilder` which is capable of handling
different value types. In these cases, we simply skip over the value
instead of encoding a `nil` value which results in a panic inside the
gob encoder.

-